requirements.txt
fastapi
uvicorn
requests
requests-oauthlib
pydantic
python-dotenv
